Decouple merge plugin from aura merge plugin
