+++ src/Nethermind/Nethermind.Merge.AuRa/AuRaMergePlugin.cs
@@ -6,7 +6,6 @@ using Nethermind.Api;
 using Nethermind.Api.Extensions;
 using Nethermind.Config;
 using Nethermind.Consensus;
-using Nethermind.Consensus.AuRa.Config;
 using Nethermind.Consensus.AuRa.InitializationSteps;
 using Nethermind.Consensus.AuRa.Transactions;
 using Nethermind.Consensus.Producers;
@@ -72,6 +71,11 @@ namespace Nethermind.Merge.AuRa
                 _manualTimestamper!,
                 _blocksConfig,
                 _api.LogManager);
+
+        protected override IBlockFinalizationManager InitializeMergeFinilizationManager()
+        {
+            return new MergeFinalizationManager(_blockFinalizationManager, _api.FinalizationManager, _poSSwitcher);
+        }
 
         public bool ShouldRunSteps(INethermindApi api)
         {
