+++ src/Nethermind/Nethermind.Merge.Plugin/MergeFinalizationManager.cs
@@ -4,19 +4,16 @@
 using System;
 using Nethermind.Blockchain;
 using Nethermind.Consensus;
-using Nethermind.Consensus.AuRa;
 using Nethermind.Consensus.Processing;
 using Nethermind.Core;
 using Nethermind.Core.Crypto;
 
 namespace Nethermind.Merge.Plugin
 {
-    public class MergeFinalizationManager : IManualBlockFinalizationManager, IAuRaBlockFinalizationManager
+    public class MergeFinalizationManager : IManualBlockFinalizationManager
     {
-        private readonly IManualBlockFinalizationManager _manualBlockFinalizationManager;
-        private readonly IAuRaBlockFinalizationManager? _auRaBlockFinalizationManager;
-        private bool HasAuRaFinalizationManager => _auRaBlockFinalizationManager is not null;
-        private bool IsPostMerge { get; set; }
+        protected readonly IManualBlockFinalizationManager _manualBlockFinalizationManager;
+        protected bool IsPostMerge { get; set; }
 
         public event EventHandler<FinalizeEventArgs>? BlocksFinalized;
 
@@ -24,7 +21,6 @@ namespace Nethermind.Merge.Plugin
             IBlockFinalizationManager? blockFinalizationManager, IPoSSwitcher poSSwitcher)
         {
             _manualBlockFinalizationManager = manualBlockFinalizationManager;
-            _auRaBlockFinalizationManager = blockFinalizationManager as IAuRaBlockFinalizationManager;
 
             poSSwitcher.TerminalBlockReached += OnSwitchHappened;
             if (poSSwitcher.HasEverReachedTerminalBlock())
@@ -33,8 +29,6 @@ namespace Nethermind.Merge.Plugin
             }
 
             _manualBlockFinalizationManager.BlocksFinalized += OnBlockFinalized;
-            if (HasAuRaFinalizationManager)
-                _auRaBlockFinalizationManager!.BlocksFinalized += OnBlockFinalized;
         }
 
         private void OnSwitchHappened(object? sender, EventArgs e)
@@ -42,7 +36,7 @@ namespace Nethermind.Merge.Plugin
             IsPostMerge = true;
         }
 
-        private void OnBlockFinalized(object? sender, FinalizeEventArgs e)
+        protected void OnBlockFinalized(object? sender, FinalizeEventArgs e)
         {
             BlocksFinalized?.Invoke(this, e);
         }
@@ -52,44 +46,9 @@ namespace Nethermind.Merge.Plugin
             _manualBlockFinalizationManager.MarkFinalized(finalizingBlock, finalizedBlock);
         }
 
-        public long GetLastLevelFinalizedBy(Hash256 blockHash)
-        {
-            if (_auRaBlockFinalizationManager is not null)
-            {
-                return _auRaBlockFinalizationManager.GetLastLevelFinalizedBy(blockHash);
-            }
-
-            throw new InvalidOperationException(
-                $"{nameof(GetLastLevelFinalizedBy)} called when empty {nameof(_auRaBlockFinalizationManager)} is null.");
-        }
-
-        public long? GetFinalizationLevel(long level)
-        {
-            if (_auRaBlockFinalizationManager is not null)
-            {
-                return _auRaBlockFinalizationManager.GetFinalizationLevel(level);
-            }
-
-            throw new InvalidOperationException(
-                $"{nameof(GetFinalizationLevel)} called when empty {nameof(_auRaBlockFinalizationManager)} is null.");
-        }
-
-        public void SetMainBlockProcessor(IBlockProcessor blockProcessor)
-        {
-            _auRaBlockFinalizationManager!.SetMainBlockProcessor(blockProcessor);
-        }
-
-        public void Dispose()
-        {
-            if (IsPostMerge && HasAuRaFinalizationManager)
-            {
-                _auRaBlockFinalizationManager!.Dispose();
-            }
-        }
-
         public Hash256 LastFinalizedHash { get => _manualBlockFinalizationManager.LastFinalizedHash; }
 
-        public long LastFinalizedBlockLevel
+        public virtual long LastFinalizedBlockLevel
         {
             get
             {
@@ -98,13 +57,10 @@ namespace Nethermind.Merge.Plugin
                     return _manualBlockFinalizationManager.LastFinalizedBlockLevel;
                 }
 
-                if (HasAuRaFinalizationManager)
-                {
-                    return _auRaBlockFinalizationManager!.LastFinalizedBlockLevel;
-                }
-
                 return 0;
             }
         }
+
+        public virtual void Dispose() { }
     }
 }
